# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DeleteResponse {
  success: boolean;
}

export interface Node {
  id: number;
  sessionId: string;
  nickname: string;
  wallet: string;
  /** @nullable */
  modelName: string | null;
  /** @nullable */
  modelNumber: string | null;
  internetSpeed: string;
  vram: string;
  /** @nullable */
  ram: string | null;
  walletHidden: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface NodeWithStats {
  id: number;
  sessionId: string;
  nickname: string;
  wallet: string;
  /** @nullable */
  modelName: string | null;
  /** @nullable */
  modelNumber: string | null;
  internetSpeed: string;
  vram: string;
  /** @nullable */
  ram: string | null;
  walletHidden: boolean;
  createdAt: string;
  updatedAt: string;
  online: boolean;
  /** @nullable */
  lastRewardAmount: number | null;
  /** @nullable */
  lastRewardTxHash: string | null;
  /** @nullable */
  lastRewardTimestamp: string | null;
  dailyAccumulated: number;
  rewardCountToday: number;
}

export interface NodeCreate {
  sessionId: string;
  nickname: string;
  wallet: string;
  /** @nullable */
  modelName?: string | null;
  /** @nullable */
  modelNumber?: string | null;
  internetSpeed: string;
  vram: string;
  /** @nullable */
  ram?: string | null;
  walletHidden?: boolean;
}

export interface NodeUpdate {
  nickname?: string;
  wallet?: string;
  /** @nullable */
  modelName?: string | null;
  /** @nullable */
  modelNumber?: string | null;
  internetSpeed?: string;
  vram?: string;
  /** @nullable */
  ram?: string | null;
  walletHidden?: boolean;
}

export interface RewardTx {
  txHash: string;
  amount: number;
  timestamp: string;
}

export interface NetworkSummary {
  totalNodes: number;
  onlineNodes: number;
  totalDailyRewards: number;
  totalRewardCountToday: number;
}

export type ListNodesParams = {
  /**
   * UTC date (YYYY-MM-DD) used to compute daily stats. Defaults to today.
   * @pattern ^\d{4}-\d{2}-\d{2}$
   */
  date?: string;
};

export type GetNetworkSummaryParams = {
  /**
   * UTC date (YYYY-MM-DD) used to compute daily stats. Defaults to today.
   * @pattern ^\d{4}-\d{2}-\d{2}$
   */
  date?: string;
};